% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{diag.RWL}
\alias{diag.RWL}
\title{Diagnostic Analysis - Local Influence}
\usage{
diag.RWL(model, mu.link = "log", sigma.link = "log", scheme = "case.weight")
}
\arguments{
\item{model}{Object of class \code{gamlss} holding the fitted model.}

\item{mu.link}{Defines the mu.link, with "identity" link as the default for the mu parameter.}

\item{sigma.link}{Defines the sigma.link, with "identity" link as the default for the sigma parameter.}

\item{scheme}{Default is "case.weight". But, can be "response".}
}
\value{
Local influence measures.
}
\description{
Diagnostics for the BP model
}
\references{
Bourguignon, M., Santos-Neto, M. and Castro, M. (2018). A new regression model for positive data. arXiv.
}
\author{
Manoel Santos-Neto \email{manoelferreira@uaest.ufcg.edu.br}
}
