% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{envelope.RWL}
\alias{envelope.RWL}
\title{Envelopes}
\usage{
envelope.RWL(
  model,
  k = 100,
  color = "grey50",
  xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile",
  font = "serif"
)
}
\arguments{
\item{model}{object of class \code{gamlss}.}

\item{k}{number of replications for envelope construction. Default is 19.}

\item{color}{a specification for the default envelope color.}

\item{xlabel}{a label for the x axis.}

\item{ylabel}{a label for the y axis.}

\item{font}{the name of a font family for x and y axis.}
}
\value{
A simulated envelope of the class BP, GA, IG, RBS and WEI3.
}
\description{
Computes simulation envelopes.
}
\references{
Atkinson, A. C. (1985) Plots, transformations and regression : an introduction to graphical methods of diagnostic regression analysis. Oxford Science Publications, Oxford.
Bourguignon, M., Santos-Neto, M. and Castro, M. (2018). A new regression model for positive data. arXiv.
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modeling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.

Santos-Neto, M., Cysneiros, F.J.A., Leiva, V., Barros, M. (2016) Reparameterized Birnbaum-Saunders
regression models with varying precision. \emph{Electronic Journal of Statistics}, 10, 2825--2855. doi: \email{10.1214/16-EJS1187}.
}
\author{
Manoel Santos-Neto \email{manoelferreira@uaest.ufcg.edu.br}
}
