% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{rrWL}
\alias{rrWL}
\title{Reparameterized Weighted Lindley Distribution}
\usage{
drwl(x,mu,phi,log.d=FALSE)
prwl(q,mu,phi,lower.tail=TRUE, log.p=FALSE)
qrwl(p, mu, phi, lower.tail=TRUE, log.q=FALSE)
rrWL(n, mu, phi)
hrwl(x, mu, phi, log.h = FALSE)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{mu, phi}{positive parameters.}

\item{x, q}{vector of quantiles.}

\item{log.d, log.p, log.q, log.h}{logical; If TRUE are given as log().}

\item{lower.tail}{logical; If TRUE, (default), P(X ≤q x) are returned, otherwise P(X > x).}

\item{p}{vector of probabilities.}
}
\description{
Density function, cumulative distribution function, quantile function, random number generation and hazard function of the reparameterized weighted Lindley distribution.
}
