% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCodes.R
\name{score_test_rwl}
\alias{score_test_rwl}
\title{Precision test}
\usage{
score_test_rwl(modelh0,modelh1)
}
\arguments{
\item{modelh0}{model under null hypothesis.}

\item{modelh1}{model under alternative hypothesis.}
}
\value{
A list with class "htest" containing the following components:

\code{statistic}	the value of the test statistic.

\code{parameter}	the degrees of freedom for the test statistic.

\code{p.value}	the p-value for the test.

\code{method}	a character string indicating what type of likelihood ratio test was performed.

\code{data.name} a character string giving the name(s) of the data
}
\description{
Tests the null hypothesis of precision fixed in RBS models against the alternative of precision variable.
}
\author{
Manoel Santos-Neto \email{manoelferreira@uaest.ufcg.edu.br}
}
